Refactor entities and remove plurals from SQL
