+++ package.json
@@ -11,12 +11,11 @@
     "lint:ci": "eslint .",
     "lint:fix": "eslint . --fix",
     "build": "npm-run-all clean tsc",
-    "pretest-podman": "podman compose up -d db-test",
-    "pretest": "docker compose up -d db-test",
+    "pretest": "f() { if [ -x $(command -v podman) ]; then podman compose up -d db-test; else docker compose up -d db-test; fi }; f",
     "test": "APP_ENV=ci jest --coverage",
     "test:ci": "APP_ENV=ci jest --ci --coverage --config=jest.config.ts",
     "check": "npm-run-all prettier:fix lint:fix test build",
-    "predev": "docker compose up -d db-dev redis",
+    "predev": "f() { if [ -x $(command -v podman) ]; then podman compose up -d db-dev redis; else docker compose up -d db-dev redis; fi }; f",
     "dev:check": "npm-run-all check dev",
     "dev": "APP_ENV=local nodemon --watch src -e ts,ejs --exec npm run start",
     "start": "npm run build && node dist/server.js | pino-colada",
