+++ src/controllers/csv-processor.ts
@@ -9,10 +9,10 @@ import { DatasetDTO } from '../dtos/dataset-dto';
 import { FileImportDTO } from '../dtos/file-import-dto';
 import { Error } from '../dtos/error';
 import { CSVHeader, ViewDTO, ViewErrDTO, ViewStream } from '../dtos/view-dto';
-import { Dataset } from '../entities/dataset';
-import { Revision } from '../entities/revision';
-import { Source } from '../entities/source';
-import { FileImport } from '../entities/file-import';
+import { Dataset } from '../entities/dataset/dataset';
+import { Revision } from '../entities/dataset/revision';
+import { Source } from '../entities/dataset/source';
+import { FileImport } from '../entities/dataset/file-import';
 import { SourceAction } from '../enums/source-action';
 import { ImportType } from '../enums/import-type';
 import { DataLocation } from '../enums/data-location';
