+++ src/entities/reference-data/category-info.ts
