+++ src/entities/reference-data/category-key-info.ts
