+++ src/entities/reference-data/category.ts
@@ -1,6 +1,6 @@
 import { Entity, BaseEntity, PrimaryColumn } from 'typeorm';
 
-@Entity({ name: 'categories' })
+@Entity({ name: 'category' })
 export class Category extends BaseEntity {
     @PrimaryColumn({ type: 'text' })
     category: string;
