+++ src/utils/sanitise-user.ts
@@ -1,6 +1,6 @@
 import { pick } from 'lodash';
 
-import { User } from '../entities/user';
+import { User } from '../entities/dataset/user';
 
 // strip anything from the user object that we do not want to expose to the client
 export const sanitiseUser = (user: User): Partial<User> => {
