+++ test/metadata-routes.test.ts
@@ -7,10 +7,10 @@ import { t, ENGLISH, WELSH } from '../src/middleware/translation';
 import { DataLakeService } from '../src/controllers/datalake';
 import { BlobStorageService } from '../src/controllers/blob-storage';
 import app, { initDb } from '../src/app';
-import { Dataset } from '../src/entities/dataset';
-import { Revision } from '../src/entities/revision';
-import { FileImport } from '../src/entities/file-import';
-import { User } from '../src/entities/user';
+import { Dataset } from '../src/entities/dataset/dataset';
+import { Revision } from '../src/entities/dataset/revision';
+import { FileImport } from '../src/entities/dataset/file-import';
+import { User } from '../src/entities/dataset/user';
 import { DatasetDTO } from '../src/dtos/dataset-dto';
 import { DimensionDTO } from '../src/dtos/dimension-dto';
 import { RevisionDTO } from '../src/dtos/revision-dto';
