+++ test/publisher-journey.test.ts
@@ -7,19 +7,19 @@ import { DataLakeService } from '../src/controllers/datalake';
 import { BlobStorageService } from '../src/controllers/blob-storage';
 import app, { initDb } from '../src/app';
 import { ENGLISH, t, WELSH } from '../src/middleware/translation';
-import { Dataset } from '../src/entities/dataset';
-import { DatasetInfo } from '../src/entities/dataset-info';
+import { Dataset } from '../src/entities/dataset/dataset';
+import { DatasetInfo } from '../src/entities/dataset/dataset-info';
 import { DatasetDTO } from '../src/dtos/dataset-dto';
 import { FileImportDTO } from '../src/dtos/file-import-dto';
 import { ViewErrDTO } from '../src/dtos/view-dto';
 import { MAX_PAGE_SIZE, MIN_PAGE_SIZE } from '../src/controllers/csv-processor';
 import DatabaseManager from '../src/db/database-manager';
-import { User } from '../src/entities/user';
-import { FileImport } from '../src/entities/file-import';
+import { User } from '../src/entities/dataset/user';
+import { FileImport } from '../src/entities/dataset/file-import';
 import { DataLocation } from '../src/enums/data-location';
 import { DimensionCreationDTO } from '../src/dtos/dimension-creation-dto';
 import { SourceType } from '../src/enums/source-type';
-import { Revision } from '../src/entities/revision';
+import { Revision } from '../src/entities/dataset/revision';
 
 import { createFullDataset, createSmallDataset } from './helpers/test-helper';
 import { getTestUser } from './helpers/get-user';
